+++ packages/react/src/components/Placeholder.test.tsx
@@ -4,7 +4,7 @@
 /* eslint-disable react/prop-types */
 import { ComponentRendering, RouteData } from '@sitecore-content-sdk/core/layout';
 import { expect } from 'chai';
-import { render } from '@testing-library/react';
+import { findByText, render } from '@testing-library/react';
 import React from 'react';
 import { spy, stub } from 'sinon';
 import {
@@ -25,12 +25,14 @@ import * as BYOCWrapper from './BYOCWrapper';
 import * as FEAASComponent from './FEaaSComponent';
 import * as FEAASWrapper from './FEaaSWrapper';
 import * as HiddenRendering from './HiddenRendering';
+import * as ErrorBoundary from './ErrorBoundary';
 import { MissingComponent, MissingComponentProps } from './MissingComponent';
 import { Placeholder } from './Placeholder';
 import { ComponentProps } from './PlaceholderCommon';
 import { SitecoreProvider } from './SitecoreProvider';
 
 const componentMap = new Map<string, React.FC>();
+const dynamicComponent = React.lazy(() => import('../test-data/test-dynamic-component'));
 
 // pass otherProps to page-content to test property cascading through the Placeholder
 
@@ -62,6 +64,7 @@ const DownloadCallout: React.FC<{
 
 componentMap.set('DownloadCallout', DownloadCallout);
 componentMap.set('Jumbotron', () => <div className="jumbotron-mock" />);
+componentMap.set('DynamicComponent', dynamicComponent);
 
 describe('<Placeholder />', () => {
   const testData = [
@@ -401,6 +404,44 @@ describe('BYOC fallback', () => {
     byocComponentStub.restore();
     byocWrapperStub.restore();
   });
+
+  it('should render ErrorBoundary without Suspense for byoc wrapper', () => {
+    const component = byocWrapperData.sitecore.route as RouteData;
+    const phKey = 'main';
+
+    byocComponentStub = stub(BYOCComponent, 'BYOCComponent').callsFake(() => (
+      <p className="byoc-component">Foo</p>
+    ));
+
+    byocWrapperStub = stub(BYOCWrapper, 'BYOCWrapper').callsFake(() => (
+      <div className="byoc-wrapper">
+        <BYOCComponent.BYOCComponent />
+      </div>
+    ));
+
+    const errorBoundarySpy = spy(ErrorBoundary, 'default');
+
+    const renderedComponent = render(
+      <SitecoreProvider componentMap={componentMap}>
+        <Placeholder name={phKey} rendering={component} />
+      </SitecoreProvider>
+    );
+
+    expect(errorBoundarySpy.calledWithMatch({ isDynamic: true })).to.be.true;
+    expect(renderedComponent.container.innerHTML).to.not.contain('Loading component...');
+
+    expect(renderedComponent.container.querySelectorAll('.byoc-wrapper').length).to.equal(1);
+
+    const components = renderedComponent.container.querySelectorAll('.byoc-component');
+
+    expect(components.length).to.equal(2);
+
+    expect(components[0].textContent).to.equal('Foo');
+    expect(components[1].textContent).to.equal('Foo');
+
+    byocComponentStub.restore();
+    byocWrapperStub.restore();
+  });
 });
 
 describe('FEaaS fallback', () => {
@@ -435,6 +476,36 @@ describe('FEaaS fallback', () => {
     feaasComponentStub.restore();
     feaasWrapperStub.restore();
   });
+});
+
+it('should render Suspense when disableSuspense is false', async () => {
+  const component = normalModeDevData.sitecore.route as RouteData;
+  const phKey = 'main';
+
+  const renderedComponent = render(
+    <SitecoreProvider componentMap={componentMap}>
+      <Placeholder name={phKey} disableSuspense={false} rendering={component} />
+    </SitecoreProvider>
+  );
+
+  expect(renderedComponent.container.innerHTML).to.contain('Loading component...');
+
+  await findByText(renderedComponent.container, 'No error');
+});
+
+it('should not render Suspense when disableSuspense is true', async () => {
+  const component = normalModeDevData.sitecore.route as RouteData;
+  const phKey = 'main';
+
+  const renderedComponent = render(
+    <SitecoreProvider componentMap={componentMap}>
+      <Placeholder name={phKey} disableSuspense={true} rendering={component} />
+    </SitecoreProvider>
+  );
+
+  expect(renderedComponent.container.innerHTML).to.not.contain('Loading component...');
+
+  await findByText(renderedComponent.container, 'No error');
 });
 
 it('should render null for unknown placeholder', () => {